+++ CHANGELOG.md
@@ -1,7 +1,6 @@
-## FIXME
+## 2019-06-03
 
-* Default Node.JS version upgraded to 10.13.0.
-* BREAKING CHANGE: Started using **async/await**, so Node.JS >=7.6.0 is require
+* BREAKING CHANGE: Started using **async/await**, so Node.JS >=7.6.0 is required. Default Node.JS version upgraded to 10.13.0.
 * FEATURE: Added Topic Report API - https://github.com/citizenos/citizenos-api/issues/5
 * DEVELOPMENT: `asyncMiddleware` added so that `async/await` can be used in Express routes.
 * Upgraded to Sequelize 5.x.